requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
rapidfuzz>=3.0.0
aiosqlite>=0.19.0
Pillow>=10.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0